API improvements (#122)
* Set specific User-Agent in API client

The API client should transmit a meaningful `User-Agent` header that
will allow us clearly identify users of this API client.

* Do not report now timestamps

Instana's API optimizes the `now` case. This optimization works best
when the `time` parameter isn't defined (this means `now`).

We should therefore let the API identify what `now` means and because of
that let all the optimizations kick in.