fix(w3c trace context): do not pass down unknown flags.
The W3C trace context specification mandates that a participant must
only send known flags downstream. See
https://www.w3.org/TR/trace-context/#other-flags: "The behavior of other
flags, such as (00000100) is not defined and is reserved for future use.
Vendors MUST set those to zero."

In particular, this commit changes the way traceparent.py handles the
flags. Instead of persisting the complete flags field from traceparent,
we specifically parse the flag(s) that we understand and keep them as
individual boolean attributes. This will also make it easier to
handle the random trace ID flag correctly when updating support to W3C
trace context level 2.

In addition we now correctly pass down the version field as 00, since
that is the traceparent version we support (instead of passing down the
incoming version value).

Signed-off-by: Bastian Krol <bastian.krol@ibm.com>