fix(w3c trace context): consider traceparent format version ff invalid
This is a follow up to commit
f283a4beb4d2b7ccb9b5aae55d076c7a5156e1d0
which added some tests for parsing newer versions. One of the tests used
format version "ff", which the W3C trace context specification
explicitly defines as invalid. This commit fixes that test and also the
regex used for parsing the traceparent header, so that a header with
version ff will be rejected as invalid.

Signed-off-by: Bastian Krol <bastian.krol@ibm.com>