Fix: remove the check for the Server header of the Instana agent.
The check for the Server header of the Instana agent can break the announce
procedure in service meshes when there is a proxy between the tracer and the
Instana agent. This happens because some proxies do not forward the original
Server header. So instead, the collector checks whether the HTTP status code
is in the 2xx range.

Signed-off-by: Paulo Vital <paulo.vital@ibm.com>