fix(agent.host): Missing spans when handling forked processes.
In multithreaded applications (services), only the first or master
process is announced to the Agent when the service starts, but if a
not announced forked process handles the (HTTP) ENTRY requests and it
contains one or more EXIT spans created during the first request,
those EXIT spans are missed and not reported.

This fix returns True when agent.host.can_send() method is called and
detects that the current process is a fork and has finished the
announcement process to the Agent with a valid FSM state.

Signed-off-by: Paulo Vital <paulo.vital@ibm.com>