fix: _log() Multiple Values for 'stacklevel'.
The error `_log() got multiple values for keyword argument 'stacklevel'`
may occur when you pass both a positional argument and a keyword
argument with the same name (`stacklevel`) to the `Logger._log()`
function.

This fix checks if the `stacklevel` is a key of the `kwargs` dictionary
before calling the `Logger._log()` function, considering it if present.

Signed-off-by: Paulo Vital <paulo.vital@ibm.com>